refactor: moves get all pets to table with loading - closes #24
